+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/ErrorManagement/Controller/Internal/ErrorManagementAction.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\ErrorManagement;
+namespace Akeneo\Connectivity\Connection\Infrastructure\ErrorManagement\Controller\Internal;
 
 use Akeneo\Connectivity\Connection\Application\ErrorManagement\Query\GetConnectionBusinessErrorsHandler;
 use Akeneo\Connectivity\Connection\Application\ErrorManagement\Query\GetConnectionBusinessErrorsQuery;
