+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/DependencyInjection/AkeneoConnectivityConnectionExtension.php
@@ -37,8 +37,8 @@ class AkeneoConnectivityConnectionExtension extends Extension
         $loader->load('Audit/queries.yml');
 
         $loader->load('ErrorManagement/commands.yml');
+        $loader->load('ErrorManagement/controllers.yml');
         $loader->load('ErrorManagement/event_subscribers.yml');
-        $loader->load('ErrorManagement/install.yml');
         $loader->load('ErrorManagement/queries.yml');
 
         $loader->load('Marketplace/controllers.yml');
