+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/ErrorManagement/controllers.yml
@@ -1,0 +1,5 @@
+services:
+    Akeneo\Connectivity\Connection\Infrastructure\ErrorManagement\Controller\Internal\ErrorManagementAction:
+        public: true
+        arguments:
+            - '@akeneo_connectivity.connection.application.error_management.handler.get_connection_business_errors'
