+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/routing.yml
@@ -18,9 +18,7 @@ akeneo_connectivity_connection_connect_apps_v1_openid_public_key:
 # Internal API
 
 akeneo_connectivity_connection_error_management_rest:
-    resource: ./routing/error-management.yml
-    name_prefix: akeneo_connectivity_connection_error_management_rest_
-    prefix: /rest/connections
+    resource: ./ErrorManagement/routing.yml
 
 akeneo_connectivity_connection_audit_rest:
     resource: ./Audit/routing.yml
@@ -67,15 +65,6 @@ akeneo_connectivity_connection_connect_apps_authorize:
 
 akeneo_connectivity_connection_connect_apps_authenticate:
     path: '/connect/apps/authenticate'
-
-## Error Management
-akeneo_connectivity_connection_error_management_dashboard:
-    path: '/connect/data-flows/business-errors'
-
-akeneo_connectivity_connection_error_management_connection_monitoring:
-    path: '/connect/connection-settings/{code}/monitoring'
-    requirements:
-        code: '[a-zA-Z0-9_]+'
 
 ## Audit
 akeneo_connectivity_connection_audit_index:
